Fixed timing issue causing CDS provisioning to fail, needed to wait longer for SSH to come up.
Issue was that SSH was up, yet the ec2 provisioning magic for copying the public key
wasn't yet complete.  So ansible would ssh in and get an auth denied since pub key wasn't yet installed.

Don't know of a 'clean' solution for this, so fell back to a 15 second wait, then check that ssh is up.